pyproject.toml
[project]
name = "event-reminder-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==3.2.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "mysql-connector-python>=9.6.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "uvicorn>=0.42.0",
]
